FixTrying to fix 'Cannot read properties of undefined'
Error likely at line 42.
Here, tagMap.get(refObj.sha) is returning undefined, and then the code is trying to access tag.data.object.sha, which throws the error because tag is undefined.  This could happen if refObj.sha is not a key in the tagMap.  To fix this, you can add a check to ensure that tag is not undefined before trying to access its properties.